second commit

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int balance=2_000_000_000;
-        int amount=500_000_000;
-        int total= balance+amount;
+        long balance=2_000_000_000;
+        long amount=500_000_000;
+        long total= balance+amount;
         System.out.println(total);
     }
 }
